KIBON-221 - Pensum for Betreuungspensum and BetreuungsmitteilungPensum changed to BigDecimal. All other Pensen (Fachstelle and Erwerbspensum) remain as Integer - class Hierarchy as described in Jira-Task for Entities, Jax and TS

+++ ebegu-rest/src/main/java/ch/dvbern/ebegu/api/dtos/JaxVerfuegungZeitabschnitt.java
@@ -70,7 +70,7 @@ public class JaxVerfuegungZeitabschnitt extends JaxAbstractDateRangedDTO {
 
 	private boolean zuSpaetEingereicht;
 
-	private int bgPensum; //min von anspruchberechtigtesPensum und betreuungspensum
+	private BigDecimal bgPensum; //min von anspruchberechtigtesPensum und betreuungspensum
 
 	private Integer einkommensjahr;
 
@@ -255,11 +255,11 @@ public class JaxVerfuegungZeitabschnitt extends JaxAbstractDateRangedDTO {
 		this.zuSpaetEingereicht = zuSpaetEingereicht;
 	}
 
-	public int getBgPensum() {
+	public BigDecimal getBgPensum() {
 		return bgPensum;
 	}
 
-	public void setBgPensum(int bgPensum) {
+	public void setBgPensum(BigDecimal bgPensum) {
 		this.bgPensum = bgPensum;
 	}
